refactor: relocate data_loading.py -> pdft_benchmarks.datasets/
- src/pdft_benchmarks/datasets/_common.py: ensure_dir helper.
- src/pdft_benchmarks/datasets/quickdraw.py: load_quickdraw + DEFAULT_QUICKDRAW_ROOT.
- src/pdft_benchmarks/datasets/div2k.py:    load_div2k    + DEFAULT_DIV2K_ROOT.
- src/pdft_benchmarks/datasets/__init__.py: load(name, **kwargs) dispatcher
  + re-exports load_div2k / load_quickdraw.

Drops the dead `_resize_batch` helper (zero callers).

Tests moved to tests/datasets/test_loaders.py with adjusted FIXTURES
path (one level deeper now).
