+++ run_div2k_8q_blocked.py
@@ -31,7 +31,7 @@ from baselines import (
     global_dct_compress,
     global_fft_compress,
 )
-from data_loading import load_div2k
+from pdft_benchmarks.datasets.div2k import load_div2k
 from run_quickdraw import _parse_args, run_dataset
 
 DATASET_NAME = "div2k_8q_blocked"
